refactor(viewer): split cellBuilderStore into zustand domain slices
cellBuilderStore.ts was 3,947 lines and ~200 interface fields in one
create() call. It is now a package of one slice per domain under
state/cellbuilder/, composed into the SAME single store, so
useCellBuilderStore.getState() returns the identical flat shape.

state/cellBuilderStore.ts stays as the import path (a re-export facade),
so every component, helper and test is untouched.

Slices (each file's header states what it owns): document, history,
selection, cells, placement, loft, systems, catalogs, detailing,
blueprints, jobs, view state, relocations, export/import. Cross-slice
calls go through the composed get(), never through a sibling import.
Supporting modules: types, state (the composed type), shared, docCodec,
loftCells, bounds, compileJobRunner, xlsxImport.

Action bodies are moved verbatim; the only edits are the create() wrapper,
the import lists, and turning the three closures over create()'s scope
(withHistory, startCompileJob, runImport) into factories bound to set/get.

Largest resulting file is 371 lines. Adds a history-slice unit test that
the split made cheap: withHistory now binds to a plain set, so the
snapshot-push rules test without a store, a DOM or the capability layer.
